Save connection config to JSON file
Read configuration from a JSON file when the main process starts and
save it, possibly with modifications, to the same file when it ends.
Add typescript interfaces for data sent over IPC.
Send connection config through renderer-init IPC event and send any
changes with main-quit so they can be added to the config before it's
saved.