Add editor toolbar, tweak more styles
Add a toolbar to Editor containing all the options already added as menu
items and a few more.
Add the ability to close the file open in the editor.
Move the check for a save path from MainApp.promptUploadCodeFile to
uploadCodeFile now that the renderer can initiate an upload.
Check for non-ASCII characters in uploaded code files as the old Dawn
does.
Change 'upload file' keybind so it doesn't conflict with Ace.
Disable user-select on most text.
Add the ability to hide and clear the console.
Tweak Modal styles.
Rearrange App.tsx.
Move Editor into the root folder. The toolbar, toolbar groups, and
buttons were originally going to be their own components, but shoving
them all in Editor.tsx doesn't make it too long.
Improve typing for ConnectionConfigModal onChange handler.