Add most missing editor toolbar icons
Add SVGs for most of the missing icons on the editor toolbar. Not sure
what to do for the keyboard controls button.
Load SVGs as assets instead of with svgr because svgr can't handle SVG
files with namespace attributes in them and we don't really need to
import them as React components anyway.
Uninstall svgr.