Discard gamepad indices when converting to inputs
Remove RobotInputDatum, which marked protos Input objects with the
source gamepad index, and just send the Input objects from renderer to
main. I'm not sure how to communicate the index information to Runtime
or if it's important at all. Omitting this may have introduced a bug but
ts is happy for now, so...